+++ app/js/services.js
@@ -4,17 +4,13 @@
  * Services module which defines app version and FBURL.
  * @module myApp/services
  */
-var app = angular.module("myApp.services", ["ngRoute", "firebase"]);
+var app = angular.module("myApp.services", ["firebase"]);
 
 // define app version
-app.value("version", "1.0.1");
+app.value("version", "1.1.0");
 
 // re-usable factory that generates the $firebaseAuth instance
-app.factory("Auth", ["$firebaseAuth", "FBURL",
-  function($firebaseAuth, FBURL) {
-      
-    var ref = new Firebase(FBURL);
-    return $firebaseAuth(ref);
-    
+app.factory("Auth", function($firebaseAuth) {
+    return $firebaseAuth();
   }
-]);
+);